Rename src/migrator/index.utils.test.ts -> src/__utils__.test.ts
